Move all tracking specific classes to tracking
Move all tracking specific classes to tracking

#17-3
