+++ kafka/src/test/java/org/axonframework/extensions/kafka/eventhandling/consumer/AsyncFetcherTest.java
@@ -1,10 +1,11 @@
 /*
- * Copyright (c) 2010-2018. Axon Framework
+ * Copyright (c) 2010-2019. Axon Framework
+ *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *    http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -22,6 +23,14 @@ import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.common.TopicPartition;
 import org.axonframework.common.AxonConfigurationException;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.Buffer;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.ConsumerUtil;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.KafkaEventMessage;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.KafkaRecordMetaData;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.KafkaTrackingToken;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.SortedKafkaMessageBuffer;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.StreamableKafkaMessageSource;
+import org.axonframework.extensions.kafka.eventhandling.consumer.tracking.TrackingRecordConverter;
 import org.axonframework.extensions.kafka.eventhandling.producer.ProducerFactory;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.*;
